directions: move everything in idunn.datasources
Mapbox's client now has its own module and all direction modules are
inheriting the same base class.
