+++ idunn/utils/default_settings.yaml
@@ -124,7 +124,6 @@ MAPBOX_DIRECTIONS_API_BASE_URL: "https://api.mapbox.com/directions/v5/mapbox"
 MAPBOX_DIRECTIONS_ACCESS_TOKEN:
 COMBIGO_API_BASE_URL: "https://maps.combigo.com/v1.1"
 COMBIGO_API_KEY:
-PUBLIC_TRANSPORTS_RESTRICT_TO_CITIES: "" # use an empty string to allow any itinerary
 HOVE_API_BASE_URL: "https://api.navitia.io/v1/journeys"
 HOVE_API_TOKEN:
